requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
setuptools==75.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.0.4
WTForms==3.2.1
